separate audioplayeractivity into audio and cast activities
