+++ app/src/main/java/de/danoeh/antennapod/activity/CastplayerActivity.java
@@ -1,0 +1,10 @@
+package de.danoeh.antennapod.activity;
+
+/**
+ * Created by domingos on 4/16/16.
+ */
+public class CastplayerActivity extends MediaplayerInfoActivity {
+    public static final String TAG = "CastPlayerActivity";
+
+
+}
